Organize smaller widgets into different files.
