+++ src/ess/reduce/widgets/_config.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
+from ipywidgets import Layout
+
+default_layout = Layout(width='80%')
+default_style = {
+    'description_width': 'auto',
+    'value_width': 'auto',
+    'button_width': 'auto',
+}
